+++ PACECloudSDK/AppKit/API/Geo/GeoAPIManager.swift
@@ -24,13 +24,16 @@ import Foundation
  */
 class GeoAPIManager {
     var speedThreshold: Double = Constants.Configuration.defaultSpeedThreshold
-    var geoAppsScope: String = Constants.Configuration.defaultGeoAppsScope
+    var geoAppsScope: String = PACECloudSDK.shared.config?.geoAppsScope ?? Constants.Configuration.defaultGeoAppsScope
 
     private let session: URLSession
     private let cloudQueue = DispatchQueue(label: "appkit-cloud-queue")
     private let apiVersion = "2021-1"
 
     private var sessionTask: URLSessionDataTask?
+
+    private var cachedCofuFeatures: [GeoAPIFeature]?
+    private var cacheCofuLastUpdatedAt: Date?
 
     private var cachedFeatures: [GeoAPIFeature]?
     private var cacheLastUpdatedAt: Date?
@@ -57,7 +60,15 @@ class GeoAPIManager {
         self.session = URLSession(configuration: configuration)
     }
 
-    func apps(for location: CLLocation, result: @escaping (Result<[GeoGasStation], GeoApiManagerError>) -> Void) {
+    func cofuGasStations(for location: CLLocation? = nil, result: @escaping (Result<[CofuGasStation], GeoApiManagerError>) -> Void) {
+        if let location = location {
+            cofuGasStations(for: location, result: result)
+        } else {
+            cofuGasStations(result: result)
+        }
+    }
+
+    private func cofuGasStations(for location: CLLocation, result: @escaping (Result<[CofuGasStation], GeoApiManagerError>) -> Void) {
         // Speed accuracy is measured in m/s.
         // Negative value means the speed property is invalid,
         // we will therefore only consider the speed, if the
@@ -69,15 +80,16 @@ class GeoAPIManager {
 
         guard let cachedFeatures = cachedFeatures, !isCacheOutdated(for: location) else {
             // Send request if cache is not available or outdated
-            fetchPolygons(for: location) { [weak self] response in
+            fetchCofuGasStations(for: location) { [weak self] response in
                 guard let self = self else {
                     result(.failure(.unknownError))
                     return
                 }
 
                 switch response {
-                case .success(let cachedFeatures):
-                    self.loadApps(for: location, with: cachedFeatures, result: result)
+                case .success:
+                    guard let cachedFeatures = self.cachedFeatures else { return }
+                    self.loadCofuStations(with: cachedFeatures, for: location, result: result)
 
                 case .failure(let error):
                     result(.failure(error))
@@ -86,13 +98,54 @@ class GeoAPIManager {
             return
         }
 
-        // Load apps from cache
-        loadApps(for: location, with: cachedFeatures, result: result)
+        // Load cofu stations from cache
+        loadCofuStations(with: cachedFeatures, for: location, result: result)
     }
 
-    private func fetchPolygons(for location: CLLocation, result: @escaping (Result<[GeoAPIFeature], GeoApiManagerError>) -> Void) {
+    private func cofuGasStations(result: @escaping (Result<[CofuGasStation], GeoApiManagerError>) -> Void) {
+        guard let cachedCofuFeatures = cachedFeatures, !isCacheOutdated() else {
+            // Send request if cache is not available or outdated
+            fetchCofuGasStations { [weak self] response in
+                guard let self = self else {
+                    result(.failure(.unknownError))
+                    return
+                }
+
+                switch response {
+                case .success:
+                    guard let cachedCofuFeatures = self.cachedCofuFeatures else { return }
+                    self.loadCofuStations(with: cachedCofuFeatures, result: result)
+
+                case .failure(let error):
+                    result(.failure(error))
+                }
+            }
+            return
+        }
+
+        // Load cofu stations from cache
+        loadCofuStations(with: cachedCofuFeatures, result: result)
+    }
+
+    func loadCofuStations(with features: [GeoAPIFeature], for location: CLLocation? = nil, result: @escaping (Result<[CofuGasStation], GeoApiManagerError>) -> Void) {
+        cloudQueue.async {
+            let cofuStations = self.retrieveCoFuGasStations(from: features)
+
+            guard let location = location else {
+                result(.success(cofuStations))
+                return
+            }
+
+            let point = [location.coordinate.longitude, location.coordinate.latitude]
+            let apps: [CofuGasStation] = cofuStations.filter { RayCasting.contains(shape: $0.polygon?.first ?? [], point: point) }
+
+            result(.success(apps))
+        }
+    }
+
+    func fetchCofuGasStations(for location: CLLocation? = nil, result: @escaping (Result<[CofuGasStation], GeoApiManagerError>) -> Void) {
         performGeoRequest { [weak self] apiResult in
-            guard let self = self else {
+            guard let unwrappedSelf = self else {
                 result(.failure(.unknownError))
                 return
             }
@@ -101,20 +154,76 @@ class GeoAPIManager {
             case .success(let apiResponse):
                 guard let features = apiResponse?.features else {
                     // If request fails reset cache to try again next time
-                    self.resetCache()
+                    self?.resetCache()
                     result(.failure(.invalidResponse))
                     return
                 }
 
-                let cachedFeatures = self.buildCachedGeoAPIResponse(with: features, for: location)
-                self.cachedFeatures = cachedFeatures
-
-                result(.success(cachedFeatures))
+                if let location = location {
+                    let relevantFeatures = unwrappedSelf.filterRelevant(features, for: location)
+                    unwrappedSelf.cachedFeatures = relevantFeatures
+                    unwrappedSelf.cacheLastUpdatedAt = Date()
+                    unwrappedSelf.cacheCenter = location
+                    let cofuStations: [CofuGasStation] = unwrappedSelf.retrieveCoFuGasStations(from: relevantFeatures)
+                    result(.success(cofuStations))
+                } else {
+                    unwrappedSelf.cachedCofuFeatures = features
+                    unwrappedSelf.cacheCofuLastUpdatedAt = Date()
+                    let cofuStations: [CofuGasStation] = unwrappedSelf.retrieveCoFuGasStations(from: features)
+                    result(.success(cofuStations))
+                }
 
             case .failure(let error):
                 result(.failure(error))
             }
         }
+    }
+
+    func retrieveCoFuGasStations(from geoFeatures: [GeoAPIFeature]) -> [CofuGasStation] {
+        return (geoFeatures.compactMap { feature in
+            var collectionFeature: GeometryCollectionsFeature?
+            var pointValue: [Double]?
+            var polygonValue: [[[Double]]]?
+
+            switch feature.geometry {
+            case .collections(let geoCollection):
+                collectionFeature = geoCollection
+
+            case .point(let pointFeature):
+                pointValue = pointFeature.coordinates
+
+            case .polygon(let polygonFeature):
+                polygonValue = polygonFeature.coordinates
+
+            case .none:
+                return nil
+            }
+
+            if let collection = collectionFeature {
+                collection.geometries?.forEach { geometry in
+                    switch geometry {
+                    case .point(let pointFeature):
+                        pointValue = pointFeature.coordinates
+
+                    case .polygon(let polygonFeature):
+                        polygonValue = polygonFeature.coordinates
+
+                    case .collections:
+                        NSLog("[GeoAPIManager] unhandled nested GeometryCollectionsFeature")
+                    }
+                }
+            }
+
+            guard let id = feature.id,
+                  let properties = feature.properties
+            else {
+                return nil
+            }
+
+            let newProperties = Dictionary(uniqueKeysWithValues: properties.map { key, value in (key, value.value) })
+
+            return CofuGasStation(id: id, coordinates: pointValue, polygon: polygonValue, properties: newProperties)
+        })
     }
 
     private func performGeoRequest(result: @escaping (Result<GeoAPIResponse?, GeoApiManagerError>) -> Void) {
@@ -166,7 +275,7 @@ class GeoAPIManager {
                     return
                 }
 
-                let decodedResponse = self?.decodeGeoAPIResonse(geoApiData: data)
+                let decodedResponse = self?.decodeGeoAPIResponse(geoApiData: data)
                 result(.success(decodedResponse))
             })
 
@@ -174,26 +283,38 @@ class GeoAPIManager {
         }
     }
 
-    private func loadApps(for location: CLLocation, with features: [GeoAPIFeature], result: @escaping ((Result<[GeoGasStation], GeoApiManagerError>) -> Void)) {
-        let point = [location.coordinate.longitude, location.coordinate.latitude]
+    private func retrievePolygon(from feature: GeoAPIFeature) -> [[[Double]]]? {
+        switch feature.geometry {
+        case .collections(let collection):
+            return retrievePolygon(from: collection)
 
-        cloudQueue.async {
-            let apps: [GeoGasStation] = features.compactMap { feature in
-                guard let id = feature.id, let apiCoordinates = feature.geometry?.coordinates else { return nil }
+        case .polygon(let polygon):
+            return polygon.coordinates
 
-                let coordinates = apiCoordinates.flatMap { $0 }
-
-                if RayCasting.contains(shape: coordinates, point: point) {
-                    return .init(id: id, apps: feature.properties?.apps ?? [])
-                }
-                return nil
-            }
-
-            result(.success(apps))
+        case .point, .none:
+            return nil
         }
     }
 
-    private func decodeGeoAPIResonse(geoApiData: Data) -> GeoAPIResponse? {
+    private func retrievePolygon(from collection: GeometryCollectionsFeature) -> [[[Double]]]? {
+        guard let geometries = collection.geometries else { return nil }
+        for geometry in geometries {
+            switch geometry {
+            case .collections(let collection):
+                return retrievePolygon(from: collection)
+
+            case .polygon(let polygon):
+                return polygon.coordinates
+
+            default:
+                break
+            }
+        }
+
+        return nil
+    }
+
+    private func decodeGeoAPIResponse(geoApiData: Data) -> GeoAPIResponse? {
         do {
             let response = try JSONDecoder().decode(GeoAPIResponse.self, from: geoApiData)
             return response
@@ -210,30 +331,29 @@ class GeoAPIManager {
 
 // MARK: - Cache
 private extension GeoAPIManager {
-    func isCacheOutdated(for location: CLLocation) -> Bool {
+    private func isCacheOutdated(for location: CLLocation? = nil) -> Bool {
+        if let location = location {
+            return isCacheOutdated(with: location)
+        } else {
+            return isCofuCacheOutdated()
+        }
+    }
+
+    private func isCofuCacheOutdated() -> Bool {
+        guard let lastUpdated = cacheCofuLastUpdatedAt else { return true }
+        return abs(lastUpdated.timeIntervalSinceNow) > cacheMaxAge
+    }
+
+    private func isCacheOutdated(with location: CLLocation) -> Bool {
         guard let lastUpdated = cacheLastUpdatedAt, let cacheCenter = cacheCenter else { return true }
         return abs(lastUpdated.timeIntervalSinceNow) > cacheMaxAge || cacheCenter.distance(from: location) > cacheRadius
     }
 
-    func buildCachedGeoAPIResponse(with features: [GeoAPIFeature], for location: CLLocation) -> [GeoAPIFeature] {
-        let filteredResponse = features.filter({ feature in
-            guard let apiCoordinates = feature.geometry?.coordinates else { return false }
-            let coordinates = apiCoordinates.flatMap { $0 }
-
-            for coordinate in coordinates {
-                guard let lon = coordinate[safe: 0], let lat = coordinate[safe: 1] else { continue }
-                let polygonEdgeLocation = CLLocation(latitude: lat, longitude: lon)
-
-                if location.distance(from: polygonEdgeLocation) > cacheRadius {
-                    return false
-                }
-            }
-
-            return true
-        })
-
-        cacheLastUpdatedAt = Date()
-        cacheCenter = location
+    func filterRelevant(_ features: [GeoAPIFeature], for location: CLLocation) -> [GeoAPIFeature] {
+        let filteredResponse = features.filter { feature in
+            guard let geometry = feature.geometry else { return false }
+            return isInRadius(geometry: geometry, location: location)
+        }
 
         return filteredResponse
     }
@@ -253,17 +373,21 @@ extension GeoAPIManager {
             return
         }
 
-        guard let cachedFeatures = cachedFeatures, !isCacheOutdated(for: location) else {
+        guard let cachedFeatures = cachedFeatures, !isCacheOutdated() else {
             // Send request if cache is not available or outdated
-            fetchPolygons(for: location) { [weak self] response in
+            fetchCofuGasStations { [weak self] response in
                 guard let self = self else {
                     completion(false)
                     return
                 }
 
                 switch response {
-                case .success(let cachedFeatures):
-                    let isAvailable = self.isAppAvailable(for: id, location: location, features: cachedFeatures)
+                case .success(let fetchedStations):
+                    var isAvailable = self.isAppAvailable(for: id, location: location, cofuStations: fetchedStations)
+                    if !isAvailable,
+                       let cachedFeatures = self.cachedFeatures {
+                        isAvailable = self.isAppAvailable(for: id, location: location, features: cachedFeatures)
+                    }
                     completion(isAvailable)
 
                 case .failure:
@@ -273,20 +397,91 @@ extension GeoAPIManager {
             return
         }
 
-        let isAvailable = isAppAvailable(for: id, location: location, features: cachedFeatures)
+        let stations = retrieveCoFuGasStations(from: cachedFeatures)
+
+        var isAvailable = isAppAvailable(for: id, location: location, cofuStations: stations)
+        if !isAvailable {
+            isAvailable = isAppAvailable(for: id, location: location, features: cachedFeatures)
+        }
         completion(isAvailable)
     }
 
+    private func isInRadius(geometry: GeometryFeature, location: CLLocation) -> Bool {
+        switch geometry {
+        case .point(let pointFeature):
+            return isInRadius(point: pointFeature, location: location)
+
+        case .polygon(let polygonFeature):
+            return isInRadius(polygon: polygonFeature, location: location)
+
+        case .collections(let collectionFeature):
+            return isInRadius(collection: collectionFeature, location: location)
+        }
+    }
+
+    private func isInRadius(point: GeometryPointFeature, location: CLLocation) -> Bool {
+        guard let lon = point.coordinates[safe: 0], let lat = point.coordinates[safe: 1] else { return false }
+        let pointLocation = CLLocation(latitude: lat, longitude: lon)
+        return location.distance(from: pointLocation) < cacheRadius
+    }
+
+    private func isInRadius(polygon: GeometryPolygonFeature, location: CLLocation) -> Bool {
+        for coordinates in polygon.coordinates {
+            for coordinate in coordinates {
+                guard let lon = coordinate[safe: 0], let lat = coordinate[safe: 1] else { continue }
+                let polygonEdgeLocation = CLLocation(latitude: lat, longitude: lon)
+
+                if location.distance(from: polygonEdgeLocation) > cacheRadius {
+                    return false
+                }
+            }
+        }
+
+        return true
+    }
+
+    private func isInRadius(collection: GeometryCollectionsFeature, location: CLLocation) -> Bool {
+        guard let geometries = collection.geometries else { return false }
+
+        for geometry in geometries {
+            switch geometry {
+            case .point(let point):
+                return isInRadius(point: point, location: location)
+
+            default:
+                continue
+            }
+        }
+
+        for geometry in geometries {
+            return isInRadius(geometry: geometry, location: location)
+        }
+
+        return false
+    }
+
+
+    private func isAppAvailable(for id: String, location: CLLocation, cofuStations: [CofuGasStation]) -> Bool {
+        guard let app = cofuStations.first(where: { $0.id == id }),
+              let coordinates = app.coordinates,
+              let lon = coordinates[safe: 0], let lat = coordinates[safe: 1] else { return false }
+
+        return location.distance(from: CLLocation(latitude: lat, longitude: lon)) <= Constants.isPoiInRangeThreshold
+    }
+
     private func isAppAvailable(for id: String, location: CLLocation, features: [GeoAPIFeature]) -> Bool {
-        guard let app = features.first(where: { $0.id == id }),
-              let coordinates = app.geometry?.coordinates?.flatMap({ $0 }) else { return false }
+        guard let app = features.first(where: { $0.id == id }) else { return false }
 
-        let isAvailable = coordinates.contains(where: { coordinate in
-            guard let lon = coordinate[safe: 0], let lat = coordinate[safe: 1] else { return false }
-            let polygonEdgeLocation = CLLocation(latitude: lat, longitude: lon)
-            return location.distance(from: polygonEdgeLocation) <= Constants.isPoiInRangeThreshold
-        })
+        switch app.geometry {
+        case .polygon(let polygonFeature):
+            return polygonFeature.coordinates.first?.contains(where: { coordinate in
+                guard let lon = coordinate[safe: 0], let lat = coordinate[safe: 1] else { return false }
+                let polygonEdgeLocation = CLLocation(latitude: lat, longitude: lon)
+                return location.distance(from: polygonEdgeLocation) <= Constants.isPoiInRangeThreshold
+            }) ?? false
 
-        return isAvailable
+        case .none, .point, .collections:
+            return false
+        }
     }
 }